# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AirshedOS API",
    "description": "Phase 3A. Spatial command center and ephemeral officer review. No causal attribution. Demo incident sharing is simulated.",
    "version": "0.7.0"
  },
  "paths": {
...
    "/api/v1/citizen-reports/{report_id}/corroborate": {
      "post": {
        "tags": [
          "Citizen evidence"
        ],
        "summary": "Corroborate",
        "operationId": "corroborate_api_v1_citizen_reports__report_id__corroborate_post",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorroborationAssessment"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportUnavailable"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/review/cases/{case_id}": {
      "get": {
        "tags": [
          "Officer review"
        ],
        "summary": "Case",
        "operationId": "case_api_v1_review_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfficerCase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/review/cases/{case_id}/review": {
      "post": {
        "tags": [
          "Officer review"
        ],
        "summary": "Review",
        "operationId": "review_api_v1_review_cases__case_id__review_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfficerCase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AirQualityForecastContext": {
        "properties": {
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "requested_horizon_hours": {
            "type": "integer",
            "enum": [
              6,
              12,
              24
            ],
            "title": "Requested Horizon Hours"
          },
          "provider": {
            "type": "string",
            "const": "google_air_quality_forecast",
            "title": "Provider",
            "default": "google_air_quality_forecast"
          },
          "provider_status": {
            "$ref": "#/components/schemas/EnvironmentalSourceStatus"
          },
          "window": {
            "$ref": "#/components/schemas/ForecastWindow"
          },
          "retrieved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieved At"
          },
          "issued_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued At"
          },
          "region_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Code"
          },
          "hourly_forecasts": {
            "items": {
              "$ref": "#/components/schemas/HourlyAirQualityForecast"
            },
            "type": "array",
            "title": "Hourly Forecasts"
          },
          "summaries": {
            "items": {
              "$ref": "#/components/schemas/ForecastSummary"
            },
            "type": "array",
            "title": "Summaries"
          },
          "current_air_quality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AirQualityObservation"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_source_status": {
            "$ref": "#/components/schemas/EnvironmentalSourceStatus"
          },
          "provenance": {
            "$ref": "#/components/schemas/EnvironmentalProvenance"
          },
          "independence_note": {
            "type": "string",
            "title": "Independence Note",
            "default": "Google current AQ and forecast AQ share a provider/model ecosystem. Forecast is an operational outlook, not an independent corroboration source; it does not change corroboration support."
          },
          "temporal_note": {
            "type": "string",
            "title": "Temporal Note",
            "default": "Hourly forecast valid times, starting next UTC hour. Retrieval time is not issuance time. Summaries describe available forecast points, not measured future outcomes."
          },
          "policy_version": {
            "type": "string",
            "const": "provider_forecast_v1",
            "title": "Policy Version",
            "default": "provider_forecast_v1"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "requested_at",
          "generated_at",
          "requested_horizon_hours",
          "provider",
          "provider_status",
          "window",
          "retrieved_at",
          "issued_at",
          "region_code",
          "hourly_forecasts",
          "summaries",
          "current_air_quality",
          "current_source_status",
          "provenance",
          "independence_note",
          "temporal_note",
          "policy_version"
        ],
        "title": "AirQualityForecastContext"
      },
...
      "AirQualityObservation": {
        "properties": {
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "retrieved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Retrieved At"
          },
          "status": {
            "type": "string",
            "enum": [
              "live",
              "cached"
            ],
            "title": "Status",
            "default": "live"
          },
          "provenance": {
            "$ref": "#/components/schemas/EnvironmentalProvenance"
          },
          "region_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Code"
          },
          "indexes": {
            "items": {
              "$ref": "#/components/schemas/AirQualityIndex"
            },
            "type": "array",
            "title": "Indexes"
          },
          "pollutants": {
            "items": {
              "$ref": "#/components/schemas/PollutantMeasurement"
            },
            "type": "array",
            "title": "Pollutants"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "observed_at",
          "retrieved_at",
          "status",
          "provenance",
          "region_code",
          "indexes",
          "pollutants"
        ],
        "title": "AirQualityObservation"
      },
...
      "CaseSnapshot": {
        "properties": {
          "record": {
            "$ref": "#/components/schemas/StructuredReport"
          },
          "assessment": {
            "$ref": "#/components/schemas/CorroborationAssessment"
          },
          "jurisdiction": {
            "$ref": "#/components/schemas/JurisdictionResolution"
          },
          "evidence_sha256": {
            "type": "string",
            "title": "Evidence Sha256"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "record",
          "assessment",
          "jurisdiction",
          "evidence_sha256"
        ],
        "title": "CaseSnapshot"
      },
...
      "CorroborationAssessment": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "event_family": {
            "$ref": "#/components/schemas/EventFamily"
          },
          "support_level": {
            "$ref": "#/components/schemas/SupportLevel"
          },
          "recommended_next_step": {
            "$ref": "#/components/schemas/NextStep"
          },
          "temporal_basis": {
            "type": "string",
            "const": "submission_time_proxy",
            "title": "Temporal Basis",
            "default": "submission_time_proxy"
          },
          "submission_time": {
            "type": "string",
            "format": "date-time",
            "title": "Submission Time"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/CorroborationRuleResult"
            },
            "type": "array",
            "title": "Rules"
          },
          "aggregation_explanation": {
            "type": "string",
            "title": "Aggregation Explanation"
          },
          "contributing_rule_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Contributing Rule Ids"
          },
          "source_summary": {
            "items": {
              "$ref": "#/components/schemas/SourceSummary"
            },
            "type": "array",
            "title": "Source Summary"
          },
          "environmental_context": {
            "$ref": "#/components/schemas/EnvironmentalContext"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "forecast_outlook": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AirQualityForecastContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "policy_version": {
            "type": "string",
            "const": "corroboration_v1",
            "title": "Policy Version",
            "default": "corroboration_v1"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "report_id",
          "generated_at",
          "event_type",
          "event_family",
          "support_level",
          "recommended_next_step",
          "temporal_basis",
          "submission_time",
          "rules",
          "aggregation_explanation",
          "contributing_rule_ids",
          "source_summary",
          "environmental_context",
          "limitations",
          "provenance",
          "forecast_outlook",
          "policy_version"
        ],
        "title": "CorroborationAssessment"
      },
      "CorroborationRuleResult": {
        "properties": {
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "inputs_used": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Inputs Used"
          },
          "evidence_references": {
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "type": "array",
            "title": "Evidence References"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rule_id",
          "source",
          "verdict",
          "summary",
          "inputs_used",
          "evidence_references",
          "generated_at"
        ],
        "title": "CorroborationRuleResult"
      },
      "EnvironmentalContext": {
        "properties": {
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "requested_reference_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Reference Time"
          },
          "time_mode": {
            "type": "string",
            "const": "current",
            "title": "Time Mode",
            "default": "current"
          },
          "temporal_note": {
            "type": "string",
            "title": "Temporal Note",
            "default": "AQ, weather and FIRMS query current conditions. Satellite uses its explicit search window."
          },
          "satellite": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SatelliteAtmosphericContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "air_quality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AirQualityObservation"
              },
              {
                "type": "null"
              }
            ]
          },
          "weather": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MeteorologicalObservation"
              },
              {
                "type": "null"
              }
            ]
          },
          "fires": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FireObservation"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fires"
          },
          "fire_dataset": {
            "type": "string",
            "title": "Fire Dataset"
          },
          "fire_search_radius_km": {
            "type": "number",
            "title": "Fire Search Radius Km"
          },
          "fire_window_days": {
            "type": "integer",
            "title": "Fire Window Days"
          },
          "source_statuses": {
            "$ref": "#/components/schemas/EnvironmentalSourceStatuses"
          },
          "incident_relationship": {
            "type": "string",
            "const": "independent_context_not_demo_corroboration",
            "title": "Incident Relationship",
            "default": "independent_context_not_demo_corroboration"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "requested_at",
          "generated_at",
          "requested_reference_time",
          "time_mode",
          "temporal_note",
          "satellite",
          "air_quality",
          "weather",
          "fires",
          "fire_dataset",
          "fire_search_radius_km",
          "fire_window_days",
          "source_statuses",
          "incident_relationship"
        ],
        "title": "EnvironmentalContext"
      },
...
      "EnvironmentalSourceStatuses": {
        "properties": {
          "air_quality": {
            "$ref": "#/components/schemas/EnvironmentalSourceStatus"
          },
          "weather": {
            "$ref": "#/components/schemas/EnvironmentalSourceStatus"
          },
          "fires": {
            "$ref": "#/components/schemas/EnvironmentalSourceStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "air_quality",
          "weather",
          "fires"
        ],
        "title": "EnvironmentalSourceStatuses"
      },
...
      "EventFamily": {
        "type": "string",
        "enum": [
          "COMBUSTION",
          "DUST",
          "ATMOSPHERIC_HAZE",
          "TRAFFIC",
          "NONE_OR_UNCERTAIN"
        ],
        "title": "EventFamily"
      },
      "EventType": {
        "type": "string",
        "enum": [
          "open_burning",
          "industrial_smoke",
          "construction_dust",
          "road_dust",
          "vehicular_emissions",
          "fire_or_combustion",
          "haze_or_smog",
          "no_visible_pollution",
          "uncertain",
          "other"
        ],
        "title": "EventType"
      },
      "EvidenceReference": {
        "properties": {
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "retrieved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Retrieved At"
          },
          "age_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Seconds"
          },
          "submission_offset_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Offset Seconds"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelProvenance"
              },
              {
                "$ref": "#/components/schemas/EnvironmentalProvenance"
              },
              {
                "$ref": "#/components/schemas/Provenance"
              }
            ],
            "title": "Provenance"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_id",
          "observed_at",
          "retrieved_at",
          "age_seconds",
          "submission_offset_seconds",
          "provenance"
        ],
        "title": "EvidenceReference"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonValue": {},
...
      "NextStep": {
        "type": "string",
        "enum": [
          "FIELD_VERIFICATION",
          "MONITOR",
          "REVIEW",
          "NO_ACTION_FROM_CURRENT_EVIDENCE"
        ],
        "title": "NextStep"
      },
      "OfficerCase": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "snapshot": {
            "$ref": "#/components/schemas/CaseSnapshot"
          },
          "review": {
            "$ref": "#/components/schemas/OfficerReview"
          },
          "storage": {
            "type": "string",
            "const": "ephemeral_process_local",
            "title": "Storage",
            "default": "ephemeral_process_local"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "created_at",
          "expires_at",
          "snapshot",
          "review",
          "storage"
        ],
        "title": "OfficerCase"
      },
      "OfficerReview": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/ReviewState",
            "default": "NEW"
          },
          "revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revision",
            "default": 0
          },
          "action_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action At"
          },
          "allowed_transitions": {
            "items": {
              "$ref": "#/components/schemas/ReviewState"
            },
            "type": "array",
            "title": "Allowed Transitions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "state",
          "revision",
          "action_at",
          "allowed_transitions"
        ],
        "title": "OfficerReview"
      },
...
      "Provenance": {
        "properties": {
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_id",
          "method",
          "is_demo",
          "note"
        ],
        "title": "Provenance"
      },
...
      "ReportUnavailable": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": "Report expired or unavailable. Analyze the image again to corroborate it."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ReportUnavailable"
      },
      "ReviewRequest": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/ReviewState"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "state",
          "expected_revision"
        ],
        "title": "ReviewRequest"
      },
      "ReviewState": {
        "type": "string",
        "enum": [
          "NEW",
          "UNDER_REVIEW",
          "ACKNOWLEDGED",
          "MONITORING",
          "CLOSED_NO_ACTION"
        ],
        "title": "ReviewState"
      },
...
      "SatelliteAtmosphericContext": {
        "properties": {
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "query_latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Query Latitude"
          },
          "query_longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Query Longitude"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "search_window": {
            "$ref": "#/components/schemas/SatelliteSearchWindow"
          },
          "retrieval_radius_km": {
            "type": "number",
            "title": "Retrieval Radius Km"
          },
          "grid_scale_m": {
            "type": "number",
            "title": "Grid Scale M"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/SatelliteProductResult"
            },
            "type": "array",
            "title": "Products"
          },
          "provider_status": {
            "$ref": "#/components/schemas/EnvironmentalSourceStatus"
          },
          "availability": {
            "type": "string",
            "enum": [
              "complete",
              "partial",
              "none"
            ],
            "title": "Availability"
          },
          "provenance": {
            "$ref": "#/components/schemas/EnvironmentalProvenance"
          },
          "temporal_note": {
            "type": "string",
            "title": "Temporal Note",
            "default": "Latest usable observation per product; acquisition times may differ. Not real-time."
          },
          "measurement_note": {
            "type": "string",
            "title": "Measurement Note",
            "default": "Satellite atmospheric columns are regional context and are not equivalent to ground-level pollutant concentrations."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "query_latitude",
          "query_longitude",
          "requested_at",
          "generated_at",
          "search_window",
          "retrieval_radius_km",
          "grid_scale_m",
          "products",
          "provider_status",
          "availability",
          "provenance",
          "temporal_note",
          "measurement_note"
        ],
        "title": "SatelliteAtmosphericContext"
      },
...
      "SourceState": {
        "type": "string",
        "enum": [
          "live",
          "cached",
          "unavailable",
          "not_configured",
          "error"
        ],
        "title": "SourceState"
      },
      "SourceSummary": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "status": {
            "$ref": "#/components/schemas/SourceState"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "status",
          "verdict",
          "message"
        ],
        "title": "SourceSummary"
      },
...
      "SupportLevel": {
        "type": "string",
        "enum": [
          "STRONG",
          "MODERATE",
          "WEAK",
          "INSUFFICIENT",
          "CONFLICTING"
        ],
        "title": "SupportLevel"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Verdict": {
        "type": "string",
        "enum": [
          "SUPPORTS",
          "WEAKLY_SUPPORTS",
          "NEUTRAL",
          "CONTRADICTS",
          "UNAVAILABLE",
          "STALE",
          "NOT_APPLICABLE"
        ],
        "title": "Verdict"
      },
...
    }
  }
}